+++ rmk/src/split/peripheral.rs
@@ -5,6 +5,11 @@ use embassy_futures::select::{Either4, select4};
 use embedded_io_async::{Read, Write};
 #[cfg(all(feature = "_ble", feature = "storage"))]
 use {super::ble::PeerAddress, crate::channel::FLASH_CHANNEL};
+#[cfg(feature = "controller")]
+use {
+    crate::channel::{CONTROLLER_CHANNEL, send_controller_event},
+    crate::event::ControllerEvent,
+};
 #[cfg(feature = "_ble")]
 use {crate::storage::Storage, embedded_storage_async::nor_flash::NorFlash, trouble_host::prelude::*};